+++ src/pytools/api/_api.py
@@ -354,7 +354,7 @@ def is_list_like(obj: Any) -> bool:
 def to_tuple(
     values: Union[Iterable[T], T],
     *,
-    element_type: Union[Type[T], Tuple[Type, ...], None] = None,
+    element_type: Optional[Union[Type[T], Tuple[Type, ...]]] = None,
     arg_name: Optional[str] = None,
 ) -> Tuple[T, ...]:
     """
@@ -366,12 +366,12 @@ def to_tuple(
       element
 
     :param values: one or more elements to return as a tuple
-    :param element_type: expected type of the values; raise a :class:`TypeError` if one
-        or more values do not implement this type
+    :param element_type: expected type of the values, or a tuple of alternative types
+        of which each value must match at least one
     :param arg_name: name of the argument as which the values were passed to a function
         or method; used when composing the :class:`TypeError` message
     :return: the values as a tuple
-    :raise TypeError: one or more values did not match the expected type
+    :raise TypeError: one or more values did not match the expected type(s)
     """
 
     return _to_collection(
@@ -386,7 +386,7 @@ def to_tuple(
 def to_list(
     values: Union[Iterable[T], T],
     *,
-    element_type: Union[Type[T], Tuple[Type, ...], None] = None,
+    element_type: Optional[Union[Type[T], Tuple[Type, ...]]] = None,
     arg_name: Optional[str] = None,
 ) -> List[T]:
     """
@@ -398,12 +398,12 @@ def to_list(
       element
 
     :param values: one or more elements to return as a list
-    :param element_type: expected type of the values; raise a :class:`TypeError` if one
-        or more values do not implement this type
+    :param element_type: expected type of the values, or a tuple of alternative types
+        of which each value must match at least one
     :param arg_name: name of the argument as which the values were passed to a function
         or method; used when composing the :class:`TypeError` message
     :return: the values as a list
-    :raise TypeError: one or more values did not match the expected type
+    :raise TypeError: one or more values did not match the expected type(s)
     """
 
     return _to_collection(
@@ -418,7 +418,7 @@ def to_list(
 def to_set(
     values: Union[Iterable[T], T],
     *,
-    element_type: Union[Type[T], Tuple[Type, ...], None] = None,
+    element_type: Optional[Union[Type[T], Tuple[Type, ...]]] = None,
     arg_name: Optional[str] = None,
 ) -> Set[T]:
     """
@@ -430,12 +430,12 @@ def to_set(
       element
 
     :param values: one or more elements to return as a set
-    :param element_type: expected type of the values; raise a :class:`TypeError` if one
-        or more values do not implement this type
+    :param element_type: expected type of the values, or a tuple of alternative types
+        of which each value must match at least one
     :param arg_name: name of the argument as which the values were passed to a function
         or method; used when composing the :class:`TypeError` message
     :return: the values as a set
-    :raise TypeError: one or more values did not match the expected type
+    :raise TypeError: one or more values did not match the expected type(s)
     """
 
     return _to_collection(
@@ -450,7 +450,7 @@ def to_set(
 def to_collection(
     values: Union[Iterable[T], T],
     *,
-    element_type: Union[Type[T], Tuple[Type, ...], None] = None,
+    element_type: Optional[Union[Type[T], Tuple[Type, ...]]] = None,
     arg_name: Optional[str] = None,
 ) -> Collection[T]:
     """
@@ -462,12 +462,12 @@ def to_collection(
       element
 
     :param values: one or more elements to return as a collection
-    :param element_type: expected type of the values; raise a :class:`TypeError` if one
-        or more values do not implement this type
+    :param element_type: expected type of the values, or a tuple of alternative types
+        of which each value must match at least one
     :param arg_name: name of the argument as which the values were passed to a function
         or method; used when composing the :class:`TypeError` message
     :return: the values as a collection
-    :raise TypeError: one or more values did not match the expected type
+    :raise TypeError: one or more values did not match the expected type(s)
     """
     return _to_collection(
         values=values,
@@ -483,8 +483,8 @@ def _to_collection(
     *,
     collection_type: Optional[Type[Collection]],
     new_collection_type: Type[T_Collection],
-    element_type: Union[Type[T], Tuple[Type, ...], None],
-    arg_name: Optional[str] = None,
+    element_type: Optional[Union[Type[T], Tuple[Type, ...]]],
+    arg_name: Optional[str],
 ) -> T_Collection:
 
     elements: T_Collection
@@ -517,7 +517,7 @@ def _to_collection(
 def validate_type(
     value: T,
     *,
-    expected_type: Union[Type, Tuple[Type, ...]],
+    expected_type: Union[Type[T], Tuple[Type, ...]],
     optional: bool = False,
     name: Optional[str] = None,
 ) -> T:
@@ -525,13 +525,14 @@ def validate_type(
     Validate that a value implements the expected type.
 
     :param value: an arbitrary object
-    :param expected_type: the type to check for
+    :param expected_type: expected type of the values, or a tuple of alternative types
+        of which the value must match at least one
     :param optional: if ``True``, accept ``None`` as a valid value (default: ``False``)
     :param name: optional name of the argument or callable with/to which the value
         was passed; use ``"arg …"`` for arguments, or the name of a callable if
         verifying positional arguments
     :return: the value passed as arg `value`
-    :raise TypeException: the value did not match the expected type
+    :raise TypeError: one or more values did not match the expected type(s)
     """
     if expected_type is object:
         return value
@@ -548,9 +549,21 @@ def validate_type(
             message_head = f"{name} requires"
         else:
             message_head = "expected"
+
+        if not isinstance(expected_type, tuple):
+            expected_type = (expected_type,)
+        if optional:
+            expected_type = (*expected_type, type(None))
+        expected_type_str = " or ".join(t.__name__ for t in expected_type)
+
+        observed_type = type(value).__name__
+
+        # noinspection SpellCheckingInspection
+        det = "an" if observed_type[0] in "aeiou" else "a"
+
         raise TypeError(
-            f"{message_head} instance of {expected_type.__name__} "
-            f"but got a {type(value).__name__}"
+            f"{message_head} an instance of {expected_type_str} "
+            f"but got {det} {observed_type}"
         )